+++ sdk/keyvault/azure-keyvault-keys/azure/keyvault/keys/crypto/aio/client.py
@@ -72,7 +72,7 @@ class CryptographyClient(AsyncKeyVaultClientBase):
         if isinstance(key, Key):
             self._key = key
             self._key_id = parse_vault_id(key.id)
-            self._allowed_ops = frozenset(self._key.key_material.key_ops)
+            self._allowed_ops = frozenset(self._key.key.key_ops)
         elif isinstance(key, str):
             self._key = None
             self._key_id = parse_vault_id(key)
@@ -115,7 +115,7 @@ class CryptographyClient(AsyncKeyVaultClientBase):
                 self._key = await self._client.get_key(
                     self._key_id.vault_endpoint, self._key_id.name, self._key_id.version, **kwargs
                 )
-                self._allowed_ops = frozenset(self._key.key_material.key_ops)
+                self._allowed_ops = frozenset(self._key.key.key_ops)
             except HttpResponseError as ex:
                 # if we got a 403, we don't have keys/get permission and won't try to get the key again
                 # (other errors may be transient)
@@ -131,10 +131,10 @@ class CryptographyClient(AsyncKeyVaultClientBase):
             if not key:
                 return None
 
-            if key.key_material.kty.lower().startswith("ec"):
-                self._internal_key = EllipticCurveKey.from_jwk(key.key_material)
+            if key.key.kty.lower().startswith("ec"):
+                self._internal_key = EllipticCurveKey.from_jwk(key.key)
             else:
-                self._internal_key = RsaKey.from_jwk(key.key_material)
+                self._internal_key = RsaKey.from_jwk(key.key)
 
         return self._internal_key