+++ sdk/keyvault/azure-keyvault-keys/azure/keyvault/keys/crypto/client.py
@@ -75,7 +75,7 @@ class CryptographyClient(KeyVaultClientBase):
         if isinstance(key, Key):
             self._key = key
             self._key_id = parse_vault_id(key.id)
-            self._allowed_ops = frozenset(self._key.key_material.key_ops)
+            self._allowed_ops = frozenset(self._key.key.key_ops)
         elif isinstance(key, six.text_type):
             self._key = None
             self._key_id = parse_vault_id(key)
@@ -120,7 +120,7 @@ class CryptographyClient(KeyVaultClientBase):
                 self._key = self._client.get_key(
                     self._key_id.vault_endpoint, self._key_id.name, self._key_id.version, **kwargs
                 )
-                self._allowed_ops = frozenset(self._key.key_material.key_ops)
+                self._allowed_ops = frozenset(self._key.key.key_ops)
             except HttpResponseError as ex:
                 # if we got a 403, we don't have keys/get permission and won't try to get the key again
                 # (other errors may be transient)
@@ -137,10 +137,10 @@ class CryptographyClient(KeyVaultClientBase):
             if not key:
                 return None
 
-            if key.key_material.kty.lower().startswith("ec"):
-                self._internal_key = EllipticCurveKey.from_jwk(key.key_material)
+            if key.key.kty.lower().startswith("ec"):
+                self._internal_key = EllipticCurveKey.from_jwk(key.key)
             else:
-                self._internal_key = RsaKey.from_jwk(key.key_material)
+                self._internal_key = RsaKey.from_jwk(key.key)
 
         return self._internal_key