+++ sdk/keyvault/azure-keyvault-keys/samples/hello_world.py
@@ -46,7 +46,7 @@ try:
     key_ops = ["encrypt", "decrypt", "sign", "verify", "wrapKey", "unwrapKey"]
     key_name = "rsaKeyName"
     rsa_key = client.create_rsa_key(key_name, size=key_size, key_operations=key_ops)
-    print("RSA Key with name '{0}' created of type '{1}'.".format(rsa_key.name, rsa_key.key_material.kty))
+    print("RSA Key with name '{0}' created of type '{1}'.".format(rsa_key.name, rsa_key.key.kty))
 
     # Let's create an Elliptic Curve key with algorithm curve type P-256.
     # if the key already exists in the Key Vault, then a new version of the key is created.
@@ -54,7 +54,7 @@ try:
     key_curve = "P-256"
     key_name = "ECKeyName"
     ec_key = client.create_ec_key(key_name, curve=key_curve)
-    print("EC Key with name '{0}' created of type '{1}'.".format(ec_key.name, ec_key.key_material.kty))
+    print("EC Key with name '{0}' created of type '{1}'.".format(ec_key.name, ec_key.key.kty))
 
     # Let's get the rsa key details using its name
     print("\n.. Get a Key by its name")